Rename service api.service to auth-service
