+++ src/app/app.component.ts
@@ -1,5 +1,5 @@
 import { Component, OnInit, ViewEncapsulation } from '@angular/core';
-import { AuthService } from './shared/services/api.service';
+import { AuthService } from './shared/services/auth-service.service';
 
 @Component({
   selector: 'app-root',
