+++ src/app/authentication-layout/username/username.component.ts
@@ -5,7 +5,7 @@ import { Observable } from 'rxjs';
 import {
   AuthService,
   AuthResponseData,
-} from 'src/app/shared/services/api.service';
+} from 'src/app/shared/services/auth-service.service';
 import { EmailResolverService } from 'src/app/shared/services/email-resolver.service';
 import { ParticleService } from 'src/app/shared/services/particle.service';
 import {
