--- src/app/main-layout/content/dashboard/dashboard.component.html
@@ -1,1 +1,0 @@
-<h3>DASHBOARD!</h3>
