+++ src/app/shared/models/user-details.model.ts
@@ -1,5 +1,6 @@
 export class UserDetails {
     constructor(
+        public user_id: number,
         public first_name: string,
         public last_name: string,
         public email: string,
