+++ src/app/shared/services/auth-guard.service.ts
@@ -5,7 +5,7 @@ import {
   Router,
   RouterStateSnapshot,
 } from '@angular/router';
-import { AuthService } from './api.service';
+import { AuthService } from './auth-service.service';
 
 @Injectable({
   providedIn: 'root',
